// app/features/common/types/event.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface Event {
  id: string;
  name: string;
  location?: {
    description?: string;
    marker?: string;
    latitude: number;
    longitude: number;
  };
  start: string;
  end?: string;
  description?: string;
}